Some initial Jenkins support for new Docker-based build
