--- etc/init.cfg
@@ -1,11 +1,0 @@
-# Init Script
-# ===========
-#
-# Generation of an init.d script for the MCL.
-
-[init]
-recipe = collective.recipe.template
-dollar = $
-input  = ${buildout:directory}/templates/mcl.sh
-mode   = 755
-output = ${buildout:parts-directory}/bin/mcl
