--- etc/logrotate.cfg
@@ -1,10 +1,0 @@
-# Log Rotation
-# ============
-#
-# Configuration file for the logrotate daemon.
-
-[logrotate]
-recipe = collective.recipe.template
-input  = ${buildout:directory}/templates/logrotate.cfg
-output = ${buildout:parts-directory}/templates/logrotate.cfg
-
