--- etc/supervisor.cfg
@@ -1,19 +1,0 @@
-# Supervisor
-# ==========
-#
-# Recipe to build/configure the Supervisor process monitor.
-# See http://supervisord.org/ for more.
-#
-# The "username" is aliased to "user" for coherence with the Zope
-# recipe.
-
-[supervisor]
-recipe   = collective.recipe.supervisor
-port     = ${ports:supervisor}
-user     = ${:username}
-username = admin
-password = admin
-programs = 
-    10 zeo                      ${buildout:bin-directory}/zeo       [fg]                    true ${users:zeo}
-    20 zope                     ${buildout:bin-directory}/zope      [console]               true ${users:zope}
-#  Pri Name     Options         Path                                Arguments               Stderr  Effective user
