--- patches/python-ldap.patch
@@ -1,11 +1,0 @@
---- setup.cfg.orig	2016-02-25 14:18:36.000000000 -0500
-+++ setup.cfg	2016-02-25 14:18:59.000000000 -0500
-@@ -1,6 +1,6 @@
- [_ldap]
--library_dirs = /usr/lib /usr/lib64 /usr/local/lib /usr/local/lib64
--include_dirs = /usr/include /usr/include/sasl /usr/local/include /usr/local/include/sasl
-+library_dirs = /usr/local/openldap2.4/lib /usr/local/openssl/lib /usr/lib /usr/lib64 /usr/local/lib /usr/local/lib64
-+include_dirs = /usr/local/openldap2.4/include /usr/local/openssl/include /usr/include /usr/include/sasl /usr/local/include /usr/local/include/sasl
- defines = HAVE_SASL HAVE_TLS HAVE_LIBLDAP_R
- extra_compile_args = 
- extra_objects =
