--- templates/apache-httpd-ssl.conf.in
@@ -1,21 +1,0 @@
-# Apache HTTPD config for SSL
-# ===========================
-#
-# Install this in $APACHE_HTTPD_HOME/conf/vhosts-ssl.
-#
-<VirtualHost ${:local-ip}:443>
-    ServerName ${:public-hostname}
-    ServerAdmin nciweb@mail.nih.gov
-    RewriteEngine On
-    SSLEngine On
-
-    # SSL
-    SSLCertificateFile ${ssl:certificate-file}
-    SSLCertificateKeyFile ${ssl:key-file}
-
-    # Onto Zope directly (bypassing Varnish cache), and within there, Plone object "mcl";
-    RewriteRule ^/(.*) http://${hosts:zope}:${ports:zope}/VirtualHostBase/https/${:public-hostname}:443/mcl/VirtualHostRoot/${:dollar}1 [L,P]
-
-    # Add secure & http-only flags to all cookies
-    Header edit Set-Cookie (.*) "$1; HTTPOnly; Secure"
-</VirtualHost>
