--- templates/apache-httpd.conf.in
@@ -1,14 +1,0 @@
-# Apache HTTPD config
-# ===================
-#
-# Install this in $APACHE_HTTPD_HOME/conf/vhosts.
-#
-<VirtualHost ${:local-ip}:80>
-    ServerName ${:public-hostname}
-    ServerAdmin nciweb@mail.nih.gov
-    RewriteEngine On
-    ProxyRequests Off
-
-    # Send everything over to https
-    RewriteRule ^(.*) https://${:public-hostname}${:dollar}1 [L,NE,R=301]
-</VirtualHost>
