--- templates/logrotate.cfg
@@ -1,23 +1,0 @@
-# logrotate.cfg
-#
-# Log file rotation.  Note that parts/templates/logrotate.cfg is generated
-# from templates/logrotate.cfg.  Are you editing the correct file?
-
-rotate 4
-weekly
-create
-compress
-delaycompress
-
-${buildout:directory}/var/log/zope*.log {
-    sharedscripts
-    postrotate
-        /bin/kill -USR2 $(cat ${buildout:directory}/var/zope.pid)
-    endscript
-}
-
-${buildout:directory}/var/log/zeo.log {
-    postrotate
-        /bin/kill -USR2 $(cat ${buildout:directory}/var/zeo.pid)
-    endscript
-}
